+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.17.0] - 2025-08-18
+
 ### Added
 
 - All events now include an `action` field with the entity of the action that triggered them.
@@ -391,8 +393,9 @@ This update features a big rewrite into a component-based API. The core concepts
 
 Initial release.
 
-[unreleased]: https://github.com/simgine/bevy_replicon/compare/v0.16.0...HEAD
-[0.16.0]: https://github.com/simgine/bevy_replicon/compare/v0.16.0...v0.16.0
+[unreleased]: https://github.com/simgine/bevy_replicon/compare/v0.17.0...HEAD
+[0.17.0]: https://github.com/simgine/bevy_replicon/compare/v0.16.0...v0.17.0
+[0.16.0]: https://github.com/simgine/bevy_replicon/compare/v0.15.3...v0.16.0
 [0.15.3]: https://github.com/simgine/bevy_replicon/compare/v0.15.2...v0.15.3
 [0.15.2]: https://github.com/simgine/bevy_replicon/compare/v0.15.1...v0.15.2
 [0.15.1]: https://github.com/simgine/bevy_replicon/compare/v0.15.0...v0.15.1